Introduce the notions of `DoubleProposalProof` and `DoubleVoteProof`
A `DoubleProposalProof` contains the two proposed macro headers plus
some signatures to verify that they originate from the same validator.

A `DoubleVoteProof` contains two proposals (either `None`, or
`Some(block_hash)`) for the same block height, same round. A
`slot_number` is used to verify that the same validator is contained in
both of the multisignatures.